typescript rewrite
